+++ examples/package-consumer/vite.config.js
@@ -1,0 +1,7 @@
+import { velodom } from "velodom/vite-plugin";
+
+export default {
+  plugins: [
+    velodom()
+  ]
+};